@@ -11,14 +11,16 @@ class GitHooks {
 
   /// create files from dart codes.
   /// [targetPath] is the absolute path
-  static void init({String targetPath}) async {
-    await Process.run('git_hooks', ['-v']).catchError((onError) async {
+  static void init({String? targetPath}) async {
+    try {
+      await Process.run('git_hooks', ['-v']);
+    } catch (error) {
       var result = await Process.run('pub', [
         'global',
         'activate',
         '--source',
         'path',
-        Utils.getOwnPath()
+        Utils.getOwnPath() ?? '',
       ]).catchError((onError) {
         print(onError);
       });
@@ -30,17 +32,17 @@ class GitHooks {
         exit(1);
       }
       await CreateHooks.copyFile(targetPath: targetPath);
-    });
+    }
   }
 
   /// unInstall git_hooks
-  static void unInstall({String path}) async {
-    await deleteFiles();
+  static Future<bool> unInstall({String? path}) {
+    return deleteFiles();
   }
 
   /// get target file path.
   /// returns the path that the git hooks points to.
-  static Future<String> getTargetFilePath({String path}) async {
+  static Future<String?> getTargetFilePath({String? path}) async {
     return CreateHooks.getTargetFilePath();
   }
 
@@ -57,7 +59,7 @@ class GitHooks {
     try {
       params.keys.forEach((userType) async {
         if (hookList[userType.toString().split('.')[1]] == type) {
-          if (!await params[userType]()) {
+          if (!await params[userType]!()) {
             exit(1);
           }
         }
